+++ src/bl64-vcs-env.bash
@@ -1,14 +1,12 @@
 #######################################
 # BashLib64 / Module / Globals / Manage Version Control System
 #
-# Version: 1.7.0
+# Version: 2.0.0
 #######################################
 
 export BL64_VCS_MODULE="$BL64_LIB_VAR_OFF"
 
 export BL64_VCS_CMD_GIT=''
-
-export BL64_VCS_ALIAS_GIT=''
 
 export BL64_VCS_SET_GIT_NO_PAGER=''
 export BL64_VCS_SET_GIT_QUIET=''