+++ src/bl64-vcs-setup.bash
@@ -1,7 +1,7 @@
 #######################################
 # BashLib64 / Module / Setup / Manage Version Control System
 #
-# Version: 1.3.0
+# Version: 2.0.0
 #######################################
 
 #######################################
@@ -22,7 +22,6 @@ function bl64_vcs_setup() {
   bl64_dbg_lib_show_function
 
   bl64_vcs_set_command &&
-    bl64_vcs_set_alias &&
     bl64_vcs_set_options &&
     BL64_VCS_MODULE="$BL64_LIB_VAR_ON"
 
@@ -48,32 +47,6 @@ function bl64_vcs_set_command() {
   case "$BL64_OS_DISTRO" in
   ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-* | ${BL64_OS_RCK}-* | ${BL64_OS_ALP}-* | ${BL64_OS_MCOS}-*)
     BL64_VCS_CMD_GIT='/usr/bin/git'
-    ;;
-  *) bl64_check_alert_unsupported ;;
-  esac
-}
-
-#######################################
-# Create command aliases for common use cases
-#
-# * Aliases are presented as regular shell variables for easy inclusion in complex commands
-# * Use the alias without quotes, otherwise the shell will interprete spaces as part of the command
-# * Warning: bootstrap function
-#
-# Arguments:
-#   None
-# Outputs:
-#   STDOUT: None
-#   STDERR: None
-# Returns:
-#   0: always ok
-#######################################
-function bl64_vcs_set_alias() {
-  bl64_dbg_lib_show_function
-  case "$BL64_OS_DISTRO" in
-  ${BL64_OS_UB}-* | ${BL64_OS_DEB}-* | ${BL64_OS_FD}-* | ${BL64_OS_CNT}-* | ${BL64_OS_RHEL}-* | ${BL64_OS_ALM}-* | ${BL64_OS_OL}-* | ${BL64_OS_RCK}-* | ${BL64_OS_ALP}-* | ${BL64_OS_MCOS}-*)
-    # shellcheck disable=SC2034
-    BL64_VCS_ALIAS_GIT="$BL64_VCS_CMD_GIT"
     ;;
   *) bl64_check_alert_unsupported ;;
   esac