+++ test/lib/batscore-setup.bash
@@ -20,8 +20,5 @@ set +o 'nounset'
 
 # Do not set/unset: 'keyword', 'noexec'
 
-# Use 6 to enable bashlib64 function debug. WARNING: tests that rely pre-recorded output fill fail due to extra debug info
-export BL64_LIB_DEBUG=0
-
 # Test-case specific variables
 export DEVBL_TEST_BASHLIB64="${TESTMANSH_PROJECT_BUILD}/bashlib64.bash"